+++ __tests__/config/schema.ts
@@ -2,19 +2,22 @@ import ajvKeywords from 'ajv-keywords';
 import { matchersWithOptions } from 'jest-json-schema';
 
 import accountSchemas from '../schemas/account.json';
+import componentSchemas from '../schemas/component.json';
 import libSchemas from '../schemas/lib.json';
 import providerSchemas from '../schemas/provider.json';
 import rpcSchemas from '../schemas/rpc.json';
 import sequencerSchemas from '../schemas/sequencer.json';
+import { isBigInt } from '../../src/utils/num';
 
-const schemas = [accountSchemas, sequencerSchemas, providerSchemas, libSchemas, rpcSchemas];
+const matcherSchemas = [accountSchemas, libSchemas, providerSchemas, rpcSchemas, sequencerSchemas];
+const schemas = [...matcherSchemas, componentSchemas];
 const jestJsonMatchers = matchersWithOptions({ schemas }, (ajv: any) => {
   // @ts-ignore
   ajv.addKeyword({
     keyword: 'isBigInt',
     type: 'object',
     validate: (_schema: any, data: any) => {
-      return typeof data === 'bigint' && data < 2n ** 64n && data >= 0n;
+      return isBigInt(data) && data < 2n ** 64n && data >= 0n;
     },
     errors: true,
   });
@@ -26,16 +29,17 @@ export const initializeMatcher = (expect: jest.Expect) => {
   expect.extend(jestJsonMatchers);
   expect.extend({
     toMatchSchemaRef(received: object, name: string) {
-      const schema = schemas.find((s) => Object.keys(s.definitions).includes(name));
+      const schema = matcherSchemas.find((s) => Object.keys(s.definitions).includes(name));
       const $ref = `${schema?.$id}#/definitions/${name}`;
       return jestJsonMatchers.toMatchSchema.call(this, received, { $ref });
     },
   });
   expect(accountSchemas).toBeValidSchema();
-  expect(sequencerSchemas).toBeValidSchema();
-  expect(providerSchemas).toBeValidSchema();
+  expect(componentSchemas).toBeValidSchema();
   expect(libSchemas).toBeValidSchema();
+  expect(providerSchemas).toBeValidSchema();
   expect(rpcSchemas).toBeValidSchema();
+  expect(sequencerSchemas).toBeValidSchema();
 };
 
 declare global {