+++ __tests__/utils/classHash.test.ts
@@ -5,6 +5,7 @@ import {
   getSelectorFromName,
 } from '../../src/utils/hash';
 import {
+  compiledC260Casm,
   compiledErc20,
   compiledHashSierra,
   compiledHashSierraCasm,
@@ -51,6 +52,13 @@ describe('Hash Tester', () => {
       const compiledClassHash = computeCompiledClassHash(compiledHashSierraCasm);
       expect(compiledClassHash).toEqual(
         '0x5c82c98f2ab111bd50293ba64bb18cf49037374783ad2486c712709c4ba0d89'
+      );
+    });
+
+    test('Hello - CompiledClassHash Cairo2.6.0 Sierra1.5.0', () => {
+      const compiledClassHash = computeCompiledClassHash(compiledC260Casm);
+      expect(compiledClassHash).toEqual(
+        '0x1725af24fbfa8050f4514651990b30e06bb9993e4e5c1051206f1bef218b1c6'
       );
     });