+++ src/channel/index.ts
@@ -1,1 +1,3 @@
-export * from './rpc_0_6';
+export * as RPC06 from './rpc_0_6';
+export * as RPC07 from './rpc_0_7';
+export * from './rpc_0_7';