+++ src/provider/rpc.ts
@@ -1,4 +1,6 @@
-import { RpcChannel } from '../channel/rpc_0_6';
+import { ProviderInterface } from './interface';
+import { LibraryError } from './errors';
+import { RpcChannel, RPC06, RPC07 } from '../channel';
 import {
   AccountInvocations,
   BigNumberish,
@@ -28,12 +30,11 @@ import {
 import { getAbiContractVersion } from '../utils/calldata/cairo';
 import { isSierra } from '../utils/contract';
 import { RPCResponseParser } from '../utils/responseParser/rpc';
-import { ProviderInterface } from './interface';
 
 export class RpcProvider implements ProviderInterface {
-  private responseParser = new RPCResponseParser();
+  private responseParser: RPCResponseParser;
 
-  public channel: RpcChannel;
+  public channel: RPC07.RpcChannel | RPC06.RpcChannel;
 
   constructor(
     optionsOrProvider?: RpcProviderOptions | ProviderInterface | RpcProvider,
@@ -51,8 +52,10 @@ export class RpcProvider implements ProviderInterface {
       optionsOrProvider.channel !== undefined
     ) {
       this.channel = optionsOrProvider.channel;
+      this.responseParser = (optionsOrProvider as any).responseParser;
     } else {
       this.channel = new RpcChannel({ ...optionsOrProvider, waitMode: false });
+      this.responseParser = new RPCResponseParser(optionsOrProvider?.feeMarginPercentage);
     }
   }
 
@@ -107,6 +110,13 @@ export class RpcProvider implements ProviderInterface {
 
   public async getBlockWithTxs(blockIdentifier?: BlockIdentifier) {
     return this.channel.getBlockWithTxs(blockIdentifier);
+  }
+
+  public async getBlockWithReceipts(blockIdentifier?: BlockIdentifier) {
+    if (this.channel instanceof RPC06.RpcChannel)
+      throw new LibraryError('Unsupported method for RPC version');
+
+    return this.channel.getBlockWithReceipts(blockIdentifier);
   }
 
   public getStateUpdate = this.getBlockStateUpdate;
@@ -182,7 +192,7 @@ export class RpcProvider implements ProviderInterface {
     // can't be named simulateTransaction because of argument conflict with account
     return this.channel
       .simulateTransaction(invocations, options)
-      .then(this.responseParser.parseSimulateTransactionResponse);
+      .then((r) => this.responseParser.parseSimulateTransactionResponse(r));
   }
 
   public async waitForTransaction(txHash: BigNumberish, options?: waitForTransactionOptions) {
@@ -284,7 +294,7 @@ export class RpcProvider implements ProviderInterface {
         ],
         { blockIdentifier, skipValidate }
       )
-      .then(this.responseParser.parseFeeEstimateResponse);
+      .then((r) => this.responseParser.parseFeeEstimateResponse(r));
   }
 
   public async getDeclareEstimateFee(
@@ -304,7 +314,7 @@ export class RpcProvider implements ProviderInterface {
         ],
         { blockIdentifier, skipValidate }
       )
-      .then(this.responseParser.parseFeeEstimateResponse);
+      .then((r) => this.responseParser.parseFeeEstimateResponse(r));
   }
 
   public async getDeployAccountEstimateFee(
@@ -324,7 +334,7 @@ export class RpcProvider implements ProviderInterface {
         ],
         { blockIdentifier, skipValidate }
       )
-      .then(this.responseParser.parseFeeEstimateResponse);
+      .then((r) => this.responseParser.parseFeeEstimateResponse(r));
   }
 
   public async getEstimateFeeBulk(
@@ -333,7 +343,7 @@ export class RpcProvider implements ProviderInterface {
   ) {
     return this.channel
       .getEstimateFee(invocations, options)
-      .then(this.responseParser.parseFeeEstimateBulkResponse);
+      .then((r) => this.responseParser.parseFeeEstimateBulkResponse(r));
   }
 
   public async invokeFunction(