+++ src/types/api/index.ts
@@ -1,7 +1,11 @@
 export * as JRPC from './jsonrpc';
-export * from './rpcspec_0_6';
+
+export * as RPCSPEC06 from './rpcspec_0_6';
+export * as RPCSPEC07 from './rpcspec_0_7';
+export * from './rpcspec_0_7';
 
 export enum ERPCVersion {
   V0_5 = '0.5',
   V0_6 = '0.6',
+  V0_7 = '0.7',
 }