+++ src/utils/provider.ts
@@ -22,7 +22,8 @@ import { ERPCVersion, ETransactionVersion } from '../types/api';
 import { isSierra } from './contract';
 import { formatSpaces } from './hash';
 import { parse, stringify } from './json';
-import { isHex, toHex } from './num';
+import { isBigInt, isHex, isNumber, toHex } from './num';
+import { isString } from './shortString';
 import { compressProgram } from './stark';
 
 /**
@@ -54,8 +55,7 @@ export function createSierraContractClass(contract: CompiledSierra): SierraContr
  * (CompiledContract or string) -> ContractClass
  */
 export function parseContract(contract: CompiledContract | string): ContractClass {
-  const parsedContract =
-    typeof contract === 'string' ? (parse(contract) as CompiledContract) : contract;
+  const parsedContract = isString(contract) ? (parse(contract) as CompiledContract) : contract;
 
   if (!isSierra(contract)) {
     return {
@@ -89,7 +89,7 @@ export const getDefaultNodeUrl = (
  * [Reference](https://github.com/starkware-libs/cairo-lang/blob/fc97bdd8322a7df043c87c371634b26c15ed6cee/src/starkware/starknet/services/api/feeder_gateway/feeder_gateway_client.py#L148-L153)
  */
 export function formatHash(hashValue: BigNumberish): string {
-  if (typeof hashValue === 'string') return hashValue;
+  if (isString(hashValue)) return hashValue;
   return toHex(hashValue);
 }
 
@@ -115,19 +115,17 @@ export class Block {
   tag: BlockIdentifier = null;
 
   private setIdentifier(__identifier: BlockIdentifier) {
-    if (typeof __identifier === 'string' && isHex(__identifier)) {
-      this.hash = __identifier;
-    } else if (typeof __identifier === 'bigint') {
+    if (isString(__identifier)) {
+      if (isHex(__identifier)) {
+        this.hash = __identifier;
+      } else if (validBlockTags.includes(__identifier as EBlockTag)) {
+        this.tag = __identifier;
+      }
+    } else if (isBigInt(__identifier)) {
       this.hash = toHex(__identifier);
-    } else if (typeof __identifier === 'number') {
+    } else if (isNumber(__identifier)) {
       this.number = __identifier;
-    } else if (
-      typeof __identifier === 'string' &&
-      validBlockTags.includes(__identifier as EBlockTag)
-    ) {
-      this.tag = __identifier;
     } else {
-      // default
       this.tag = EBlockTag.PENDING;
     }
   }